Fix image encoding/decoding bugs
Images are stored in VImage CBaaS values as `ModelImage (Image PixelRGBA8)`.

When an image is encoded in JSON, it is a base_64 encoded bytestring that could have come from any image type (.jpeg, .tiff, .png).

When decoding an image, JuicyPixels decodeImage tries to import it into a DynamicImage, and convertRGBA8 turns it into an (Image PixelRGBA8).

The frontend javascript and python worker base library got updates to reflect this.

Also tweaked the python worker library to reconnect automatically when the connection dies.